+++ src/CleanRegex/Internal/Model/LightweightGroupAware.php
@@ -2,7 +2,6 @@
 namespace TRegx\CleanRegex\Internal\Model;
 
 use TRegx\CleanRegex\Internal\Definition;
-use TRegx\SafeRegex\preg;
 
 class LightweightGroupAware implements GroupAware
 {
@@ -29,7 +28,7 @@ class LightweightGroupAware implements GroupAware
     private function matches(): array
     {
         if ($this->matches === null) {
-            preg::match_all($this->definition->pattern, '', $this->matches);
+            \preg_match_all($this->definition->pattern, '', $this->matches);
         }
         return $this->matches;
     }